neural-agency/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.40.0
python-dotenv==1.0.1
aiosqlite==0.20.0
pydantic==2.10.3
jinja2==3.1.4
python-multipart==0.0.20
httpx==0.28.1
